requirements.txt
gunicorn
flask
ccxt
requests
decimal
python-dotenv